app 모듈 코드 정리
