+++ app/src/main/java/camp/nextstep/edu/github/MainViewModel.kt
@@ -1,0 +1,16 @@
+package camp.nextstep.edu.github
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.liveData
+import camp.nextstep.edu.github.data.GithubRepository
+import camp.nextstep.edu.github.data.GithubService
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class MainViewModel @Inject constructor(
+    private val githubService: GithubService
+) : ViewModel() {
+    val repos: LiveData<List<GithubRepository>> = liveData { emit(githubService.getRepositories()) }
+}
